Treat undefined reader dispatch macros as single line comments
